fade animation works across routes
