--- src/components/pages/profile/Container.tsx
@@ -1,10 +1,0 @@
-import { Link } from "react-router-dom";
-
-const ProfilePage = (): JSX.Element => {
-  return (<>
-    <h2>profile</h2>
-    <Link to="/">index page</Link>
-    </>);
-}
-
-export default ProfilePage;
